// src/app/types/Styles.ts
export type Breakpoints = 'xs' | 'sm' | 'md' | 'lg' | 'xl' | 'xxl'

export type Colors =
	| 'primary'
	| 'secondary'
	| 'success'
	| 'danger'
	| 'warning'
	| 'info'
	| 'dark'
	| 'light'
	| string

export type Placements =
	| 'auto'
	| 'auto-start'
	| 'auto-end'
	| 'top-end'
	| 'top'
	| 'top-start'
	| 'bottom-end'
	| 'bottom'
	| 'bottom-start'
	| 'right-start'
	| 'right'
	| 'right-end'
	| 'left-start'
	| 'left'
	| 'left-end'
	| undefined

export type Shapes =
	| 'rounded-none'
	| 'rounded-sm'
	| 'rounded'
	| 'rounded-md'
	| 'rounded-lg'
	| 'rounded-xl'
	| 'rounded-2xl'
	| 'rounded-3xl'
	| 'rounded-full'
	| 'rounded-t'
	| 'rounded-b'
	| 'rounded-r'
	| 'rounded-l'
	| string

export type Sizes =
	| 'sm'
	| 'md'
	| 'lg'
	| string

export type Triggers = 'hover' | 'focus' | 'click'